pyproject.toml
[project]
name = "dr-clivi"
version = "0.1.0"
description = "Dr. Clivi - Multi-Agent Healthcare Assistant for Diabetes and Obesity flows via WhatsApp"
authors = [{ name = "Clivi Team", email = "dev@clivi.com" }]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Google Cloud AI Platform with ADK
    "google-cloud-aiplatform[adk,agent-engines] >= 1.93.0",
    "google-genai >= 1.9.0",
    "google-adk >= 1.5.0",
    # A2A Protocol for agent communication
    "a2a-sdk >= 0.2.4",
    # Core dependencies
    "pydantic >= 2.0.0",
    "pydantic-settings >= 2.0.0",
    "python-dotenv >= 1.0.1",
    "rich >= 13.9.4",
    # WhatsApp/Messaging & Webhook server
    "fastapi >= 0.100.0",
    "uvicorn >= 0.20.0",
    "httpx >= 0.24.0",
    "requests >= 2.32.3",
    # Data handling
    "pandas >= 2.2.3",
    # Healthcare/Medical specific
    "python-dateutil >= 2.9.0",
    # Testing
    "pytest >= 7.4.4",
    "pytest-asyncio >= 0.21.1",
]

[project.optional-dependencies]
dev = [
    "black >= 25.1.0",
    "jupyter >= 1.1.1",
    "ipykernel >= 6.29.5",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["dr_clivi*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "."]
python_files = ["test_*.py", "*_test.py"]
